+++ dst/app/middlewares/errorHandler.js
@@ -3,11 +3,11 @@ Object.defineProperty(exports, "__esModule", { value: true });
 /**
  * エラーハンドラーミドルウェア
  */
-const ttts = require("@tokyotower/domain");
+const alverca = require("@alverca/domain");
 const createDebug = require("debug");
 const http_status_1 = require("http-status");
 const api_1 = require("../error/api");
-const debug = createDebug('ttts-api:middlewares:errorHandler');
+const debug = createDebug('@alverca/api:middlewares:errorHandler');
 exports.default = (err, __, res, next) => {
     debug('handling err...', err);
     if (res.headersSent) {
@@ -21,14 +21,15 @@ exports.default = (err, __, res, next) => {
     else {
         // エラー配列が入ってくることもある
         if (Array.isArray(err)) {
-            apiError = new api_1.APIError(tttsError2httpStatusCode(err[0]), err);
+            apiError = new api_1.APIError(alvercaError2httpStatusCode(err[0]), err);
         }
-        else if (err instanceof ttts.factory.errors.TTTS) {
-            apiError = new api_1.APIError(tttsError2httpStatusCode(err), [err]);
+        else if (err instanceof alverca.factory.errors.SmartTheater) {
+            apiError = new api_1.APIError(alvercaError2httpStatusCode(err), [err]);
         }
         else {
             // 500
-            apiError = new api_1.APIError(http_status_1.INTERNAL_SERVER_ERROR, [new ttts.factory.errors.TTTS('InternalServerError', err.message)]);
+            apiError =
+                new api_1.APIError(http_status_1.INTERNAL_SERVER_ERROR, [new alverca.factory.errors.SmartTheater('InternalServerError', err.message)]);
         }
     }
     res.status(apiError.code)
@@ -37,39 +38,37 @@ exports.default = (err, __, res, next) => {
     });
 };
 /**
- * TTTSエラーをHTTPステータスコードへ変換する
- * @function
- * @param {ttts.factory.errors.TTTS} err TTTSエラー
+ * AlvercaエラーをHTTPステータスコードへ変換する
  */
-function tttsError2httpStatusCode(err) {
+function alvercaError2httpStatusCode(err) {
     let statusCode = http_status_1.BAD_REQUEST;
     switch (true) {
         // 401
-        case (err instanceof ttts.factory.errors.Unauthorized):
+        case (err instanceof alverca.factory.errors.Unauthorized):
             statusCode = http_status_1.UNAUTHORIZED;
             break;
         // 403
-        case (err instanceof ttts.factory.errors.Forbidden):
+        case (err instanceof alverca.factory.errors.Forbidden):
             statusCode = http_status_1.FORBIDDEN;
             break;
         // 404
-        case (err instanceof ttts.factory.errors.NotFound):
+        case (err instanceof alverca.factory.errors.NotFound):
             statusCode = http_status_1.NOT_FOUND;
             break;
         // 409
-        case (err instanceof ttts.factory.errors.AlreadyInUse):
+        case (err instanceof alverca.factory.errors.AlreadyInUse):
             statusCode = http_status_1.CONFLICT;
             break;
         // 429
-        case (err instanceof ttts.factory.errors.RateLimitExceeded):
+        case (err instanceof alverca.factory.errors.RateLimitExceeded):
             statusCode = http_status_1.TOO_MANY_REQUESTS;
             break;
         // 502
-        case (err instanceof ttts.factory.errors.NotImplemented):
+        case (err instanceof alverca.factory.errors.NotImplemented):
             statusCode = http_status_1.NOT_IMPLEMENTED;
             break;
         // 503
-        case (err instanceof ttts.factory.errors.ServiceUnavailable):
+        case (err instanceof alverca.factory.errors.ServiceUnavailable):
             statusCode = http_status_1.SERVICE_UNAVAILABLE;
             break;
         // 400