+++ CHANGELOG.md
@@ -10,9 +10,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+- @alverca/domainで再構築
+
 ### Deprecated
 
 ### Removed
+
+- パフォーマンスルーターを削除
 
 ### Fixed